+++ cardano-api/src/Cardano/Api/Shelley.hs
@@ -228,6 +228,15 @@ module Cardano.Api.Shelley
     AcquiringFailure(..),
     SystemStart(..),
 
+    -- ** Governance
+    GovernancePoll (..),
+    GovernancePollAnswer (..),
+    GovernancePollWitness (..),
+    GovernancePollError (..),
+    renderGovernancePollError,
+    hashGovernancePoll,
+    signPollAnswerWith,
+    verifyPollAnswer,
 
     -- ** Various calculations
     LeadershipError(..),
@@ -249,6 +258,7 @@ import           Cardano.Api.Block
 import           Cardano.Api.Certificate
 import           Cardano.Api.Eras
 import           Cardano.Api.Genesis
+import           Cardano.Api.Governance.Poll
 import           Cardano.Api.InMode
 import           Cardano.Api.IPC
 import           Cardano.Api.Keys.Byron